fixes workspaces creation
