+++ main/core/Resources/modules/tools/workspaces/store/index.js
@@ -1,8 +1,10 @@
 
+import {actions} from '#/main/core/tools/workspaces/store/actions'
 import {reducer} from '#/main/core/tools/workspaces/store/reducer'
 import {selectors} from '#/main/core/tools/workspaces/store/selectors'
 
 export {
+  actions,
   reducer,
   selectors
 }
