--- main/core/Resources/modules/workspace/creation/containers/creation.jsx
@@ -1,12 +1,0 @@
-import {connect} from 'react-redux'
-
-import {CreationForm} from '#/main/core/workspace/creation/components/creation'
-
-const ConnectedWorkspaceCreationContainer = connect(
-  null,
-  null
-)(CreationForm)
-
-export {
-  ConnectedWorkspaceCreationContainer as WorkspaceCreation
-}
