--- main/core/Resources/modules/workspace/creation/store/index.js
@@ -1,8 +1,0 @@
-
-import {actions} from '#/main/core/workspace/creation/store/actions'
-import {reducer} from '#/main/core/workspace/creation/store/reducer'
-
-export {
-  actions,
-  reducer
-}
